+++ jest.config.js
@@ -4,6 +4,9 @@ const { compilerOptions } = require('./tsconfig');
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'jsdom',
+  transform: {
+    '^.+\\.tsx?$': 'esbuild-jest',
+  },
   moduleNameMapper: {
     '\\.(css|less)$': '<rootDir>/src/util/jest_css_placeholder.tsx',
     ...pathsToModuleNameMapper(compilerOptions.paths, { prefix: "<rootDir>/src" })