+++ docs/changelog.md
@@ -15,6 +15,7 @@
 - Prevent new document creation as server setting
 - Add readyForEdits flag to fix cursor jumps
 - Fix infinite loop in link-under focus shortcut
+- Add additional tests + fix shift tab into parent item
 
 # Version 3.6.0: Add omnibar 
 - Refactor to separate omnibar and search